creates plot of known scaffolds
This script takes the ygs.pl and cnv-seq output and uses them to create a plot showing the trends across X, Y & autosome. We got three nice clusters so I'm curious to hear about yours. Each window is a dot color-coded by where the scaffold it's drawn from has been predicted to map in the genome (based on comparative genomic alignment). For us, this showed that the methods work in discriminating X from Y from autosome.  You also need one more input, which is a table telling it which scaffold should be which color. Hopefully this is all explained in the comments.